Create youtube_search_shorts
The youtube_search_shorts function uses Selenium and BeautifulSoup libraries to search for shorts (YouTube's short-form videos) based on a given query.
It opens a Chrome WebDriver and navigates to the YouTube search results page corresponding to the query.
It retrieves the HTML content of the page and then closes the browser.
The HTML content is parsed using BeautifulSoup to extract the URLs of the shorts.
The function adds the prefix "youtube.com/shorts/" to each short URL.
It returns a list of the shorts' URLs.
The script can be run from the command line, allowing the user to specify the search term using the --q argument. The default search term is "Google."
The generated code includes the necessary imports, the definition of the youtube_search_shorts function, and the code to parse the command-line arguments and execute the search.
Please make sure to replace "path/to/chromedriver" with the correct path to your ChromeDriver executable.